# Conflicts:
#	pters_project/tasks/views.py
#	pters_project/trainee/templates/popup/trainee_popup_calendar_plan_reserve.html

+++ pters_project/trainee/views.py
@@ -470,7 +470,9 @@ def add_trainee_schedule_logic(request):
     # time_duration = request.POST.get('time_duration', '')
     training_time = request.POST.get('training_time', '')
     # class_type_name = request.session.get('class_type_name', '')
+    func_setting_data_update(request, 'trainee')
     setting_week_start_date = request.session.get('setting_week_start_date', 'SUN')
+    setting_single_lecture_duplicate = request.session.get('setting_single_lecture_duplicate', UN_USE)
     error = None
     class_info = None
     start_date = None
@@ -667,7 +669,8 @@ def add_trainee_schedule_logic(request):
 
     if error is None:
         schedule_result = pt_add_logic_func(training_date, start_date, end_date, request.user.id, member_ticket_id,
-                                            class_id, request, lecture_info, lecture_schedule_id)
+                                            class_id, request, lecture_info, lecture_schedule_id,
+                                            setting_single_lecture_duplicate)
         error = schedule_result['error']
         context['schedule_id'] = schedule_result['schedule_id']
     # if error is None:
@@ -898,6 +901,8 @@ class GetTraineeScheduleView(LoginRequiredMixin, AccessTestMixin, TemplateView):
         class_id = self.request.session.get('class_id', '')
         trainer_id = self.request.session.get('trainer_id', '')
         class_hour = self.request.session.get('class_hour', '')
+        func_setting_data_update(self.request, 'trainee')
+        setting_single_lecture_duplicate = self.request.session.get('setting_single_lecture_duplicate', UN_USE)
         today = datetime.date.today()
         error = None
         if date != '':
@@ -906,6 +911,10 @@ class GetTraineeScheduleView(LoginRequiredMixin, AccessTestMixin, TemplateView):
             day = 46
         start_date = today - datetime.timedelta(days=int(day))
         end_date = today + datetime.timedelta(days=int(day))
+
+        if setting_single_lecture_duplicate == UN_USE:
+            one_to_one_lecture_id = None
+
         if class_id is not None and class_id != '':
             if trainer_id == '' or trainer_id is None:
                 try:
@@ -1514,8 +1523,8 @@ class AlarmViewAjax(LoginRequiredMixin, AccessTestMixin, View):
         return render(request, self.template_name, context)
 
 
-def pt_add_logic_func(schedule_date, start_date, end_date, user_id,
-                      member_ticket_id, class_id, request, lecture_info, lecture_schedule_id):
+def pt_add_logic_func(schedule_date, start_date, end_date, user_id, member_ticket_id, class_id, request, lecture_info,
+                      lecture_schedule_id, setting_single_lecture_duplicate):
 
     class_type_name = request.session.get('class_type_name', '')
     error = None
@@ -1563,7 +1572,16 @@ def pt_add_logic_func(schedule_date, start_date, end_date, user_id,
                 note = lecture_schedule_info.note
             except ObjectDoesNotExist:
                 error = '회원 정보를 불러오지 못했습니다.[0]'
-        # else:
+        else:
+            if setting_single_lecture_duplicate == USE:
+                single_lecture_schedule_num = ScheduleTb.objects.filter(lecture_tb_id=lecture_info.lecture_id,
+                                                                        start_dt=start_date, end_dt=end_date,
+                                                                        use=USE).count()
+                if single_lecture_schedule_num == 0:
+                    schedule_duplication = SCHEDULE_DUPLICATION_ENABLE
+                else:
+                    error = '일정이 중복되었습니다.'
+
         #     lecture_info = func_get_member_ticket_one_to_one_lecture_info(class_id, member_ticket_id)
         #     lecture_schedule_id = None
 
@@ -2522,39 +2540,42 @@ def update_trainee_setting_push_logic(request):
     error = update_alarm_setting_data(class_id, request.user.id, setting_type_cd_data, setting_info_data)
 
     now = timezone.now()
+    if class_id is None or class_id == '':
+        error = '지점 정보를 불러오지 못했습니다.'
 
-    schedule_alarm_data = ScheduleAlarmTb.objects.select_related(
-        'schedule_tb').filter(class_tb_id=class_id, alarm_dt__gte=now, member_id=request.user.id, use=USE)
+    if error is None:
+        schedule_alarm_data = ScheduleAlarmTb.objects.select_related(
+            'schedule_tb').filter(class_tb_id=class_id, alarm_dt__gte=now, member_id=request.user.id, use=USE)
 
-    if setting_schedule_alarm_minute == '-1':
-        schedule_alarm_data.delete()
-    else:
-        alarm_time = now + datetime.timedelta(minutes=int(setting_schedule_alarm_minute))
-        alarm_time = alarm_time.strftime('%Y-%m-%d %H:%M:00')
+        if setting_schedule_alarm_minute == '-1':
+            schedule_alarm_data.delete()
+        else:
+            alarm_time = now + datetime.timedelta(minutes=int(setting_schedule_alarm_minute))
+            alarm_time = alarm_time.strftime('%Y-%m-%d %H:%M:00')
 
-        schedule_data = ScheduleTb.objects.select_related(
-            'member_ticket_tb').filter(class_tb_id=class_id, start_dt__gte=alarm_time,
-                                       member_ticket_tb__member_id=request.user.id,
-                                       permission_state_cd=PERMISSION_STATE_CD_APPROVE,
-                                       use=USE)
-        for schedule_info in schedule_data:
-            alarm_dt = schedule_info.start_dt - datetime.timedelta(minutes=int(setting_schedule_alarm_minute))
+            schedule_data = ScheduleTb.objects.select_related(
+                'member_ticket_tb').filter(class_tb_id=class_id, start_dt__gte=alarm_time,
+                                           member_ticket_tb__member_id=request.user.id,
+                                           permission_state_cd=PERMISSION_STATE_CD_APPROVE,
+                                           use=USE)
+            for schedule_info in schedule_data:
+                alarm_dt = schedule_info.start_dt - datetime.timedelta(minutes=int(setting_schedule_alarm_minute))
 
-            new_reg_test = True
-            for schedule_alarm_info in schedule_alarm_data:
-                if str(schedule_alarm_info.schedule_tb_id) == str(schedule_info.schedule_id):
-                    schedule_alarm_info.alarm_dt = alarm_dt
-                    schedule_alarm_info.alarm_minute = setting_schedule_alarm_minute
+                new_reg_test = True
+                for schedule_alarm_info in schedule_alarm_data:
+                    if str(schedule_alarm_info.schedule_tb_id) == str(schedule_info.schedule_id):
+                        schedule_alarm_info.alarm_dt = alarm_dt
+                        schedule_alarm_info.alarm_minute = setting_schedule_alarm_minute
+                        schedule_alarm_info.save()
+                        new_reg_test = False
+                        break
+
+                if new_reg_test:
+                    schedule_alarm_info = ScheduleAlarmTb(class_tb_id=class_id, schedule_tb_id=schedule_info.schedule_id,
+                                                          alarm_dt=alarm_dt, member_id=request.user.id,
+                                                          alarm_minute=setting_schedule_alarm_minute,
+                                                          use=USE)
                     schedule_alarm_info.save()
-                    new_reg_test = False
-                    break
-
-            if new_reg_test:
-                schedule_alarm_info = ScheduleAlarmTb(class_tb_id=class_id, schedule_tb_id=schedule_info.schedule_id,
-                                                      alarm_dt=alarm_dt, member_id=request.user.id,
-                                                      alarm_minute=setting_schedule_alarm_minute,
-                                                      use=USE)
-                schedule_alarm_info.save()
 
     if error is None:
         request.session['setting_push_from_trainer_lesson_alarm'] = int(setting_push_from_trainer_lesson_alarm)
